certificador/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
argcomplete==3.2.3
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
bitarray==2.9.2
boto3==1.34.96
botocore==1.34.96
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cytoolz==0.12.3
Django==5.0.4
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
environ==1.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.1
eth-utils==4.1.0
eth_abi==5.1.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
idna==3.7
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
lru-dict==1.2.0
multidict==6.0.5
packaging==24.0
parsimonious==0.10.0
pipx==1.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psycopg2==2.9.9
py_essentials==1.4.12
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pyunormalize==15.1.0
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
toolz==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.2
vine==5.1.0
wcwidth==0.2.13
web3==6.17.2
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
